requirements.txt
# API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.0
pydantic-settings==2.12.0

# LLM & AI
openai==1.54.4
anthropic==0.39.0
langchain==0.3.15
langchain-openai==0.2.15
langchain-anthropic==0.3.5
langchain-core==0.3.30

# Vector Store & Embeddings
faiss-cpu==1.9.0
chromadb==0.5.20
sentence-transformers==3.3.1

# Data Processing
numpy==2.1.3
pandas==2.2.3

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Logging
loguru==0.7.3
